Fix second batch of compile errors
- GameServerUtils: split TrBase assignment into TrBase0/1/2 floats
- ServerConsoleCommands: fix SV_Shutdown caller, Transmit arg order,
  remove ref from Decode (array mutated in-place, ref not needed)
- CGameAtmospheric: remove non-existent EmitParams.remainingLifetime
- AudioSystem: add S_StartLocalSound(string, int) overload for path-based calls
- CGameServerCmds: cast float yaw to int for OnDamageIndicator; replace
  direct event invocation with CGameEntities.RaiseEntityEvent
- CGameEntities: add RaiseEntityEvent / RaiseExplosion raise helpers
- CGameSound: replace removed _audio field with AudioSystem static call
- CinematicSystem, ClientSystem: add missing using ET.Core
- UISystem: add RaiseMenuDraw helper; fix UIShared call that crossed class boundary

+++ UnityProject/Assets/Scripts/Server/GameCommands.cs
@@ -212,14 +212,14 @@ namespace ET.Server
             {
                 case "health":
                     ent.Health = ent.MaxHealth;
-                    ent.Client.PS.Stats[GameConst.STAT_HEALTH] = ent.MaxHealth;
+                    ent.Client.PS.Stats[GameConst.STAT_HEALTH] = (short)ent.MaxHealth;
                     break;
                 case "ammo":
                     WeaponSystem.GiveAmmo(ent.Client.PS, -1, true);
                     break;
                 case "all":
                     ent.Health = ent.MaxHealth;
-                    ent.Client.PS.Stats[GameConst.STAT_HEALTH] = ent.MaxHealth;
+                    ent.Client.PS.Stats[GameConst.STAT_HEALTH] = (short)ent.MaxHealth;
                     WeaponSystem.GiveAmmo(ent.Client.PS, -1, true);
                     break;
                 default:
